backend/requirements.txt
asttokens==3.0.0
attrs==25.3.0
autobahn==19.11.2
Automat==25.4.16
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
comm==0.2.3
constantly==23.10.4
cryptography==45.0.5
debugpy==1.8.16
decorator==5.2.1
dhanhq==2.0.2
executing==2.2.1
Flask==3.1.1
flask-cors==6.0.1
gunicorn==22.0.0
hiredis==3.2.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
kiteconnect==5.0.1
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mibian==0.1.3
nest-asyncio==1.6.0
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.5
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
Pygments==2.19.2
pyOpenSSL==25.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pyzmq==27.0.2
redis==6.4.0
requests==2.32.4
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.2
traitlets==5.14.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
zope.interface==7.2
